Renamed utils.js to prepare.js
